requirements.txt
gradio>=4.16.0
PyPDF2>=3.0.0
groq>=0.9.0
scikit-learn>=1.3.0
numpy>=1.24.0,<2.0.0
sentence-transformers>=2.3.0
python-docx>=1.1.0